requirements.txt
Django==4.0.2
dj-database-url==0.5.0
djangorestframework==3.13.1
django-oauth-toolkit==1.7.0
drf-yasg==1.20.0
geopy==2.2.0
